+++ mllib/src/main/scala/org/apache/spark/ml/feature/ChiSqSelector.scala
@@ -17,122 +17,15 @@
 
 package org.apache.spark.ml.feature
 
-import scala.collection.mutable.ArrayBuilder
-
 import org.apache.hadoop.fs.Path
 
 import org.apache.spark.annotation.Since
-import org.apache.spark.ml._
-import org.apache.spark.ml.attribute._
-import org.apache.spark.ml.linalg._
 import org.apache.spark.ml.param._
-import org.apache.spark.ml.param.shared._
 import org.apache.spark.ml.stat.ChiSquareTest
 import org.apache.spark.ml.util._
-import org.apache.spark.sql._
-import org.apache.spark.sql.functions._
-import org.apache.spark.sql.types.{StructField, StructType}
+import org.apache.spark.sql.{DataFrame, Dataset}
+import org.apache.spark.sql.types.StructType
 
-/**
- * Params for [[ChiSqSelector]] and [[ChiSqSelectorModel]].
- */
-private[feature] trait ChiSqSelectorParams extends Params
-  with HasFeaturesCol with HasOutputCol with HasLabelCol {
-
-  /**
-   * Number of features that selector will select, ordered by ascending p-value. If the
-   * number of features is less than numTopFeatures, then this will select all features.
-   * Only applicable when selectorType = "numTopFeatures".
-   * The default value of numTopFeatures is 50.
-   *
-   * @group param
-   */
-  @Since("1.6.0")
-  final val numTopFeatures = new IntParam(this, "numTopFeatures",
-    "Number of features that selector will select, ordered by ascending p-value. If the" +
-      " number of features is < numTopFeatures, then this will select all features.",
-    ParamValidators.gtEq(1))
-  setDefault(numTopFeatures -> 50)
-
-  /** @group getParam */
-  @Since("1.6.0")
-  def getNumTopFeatures: Int = $(numTopFeatures)
-
-  /**
-   * Percentile of features that selector will select, ordered by ascending p-value.
-   * Only applicable when selectorType = "percentile".
-   * Default value is 0.1.
-   * @group param
-   */
-  @Since("2.1.0")
-  final val percentile = new DoubleParam(this, "percentile",
-    "Percentile of features that selector will select, ordered by ascending p-value.",
-    ParamValidators.inRange(0, 1))
-  setDefault(percentile -> 0.1)
-
-  /** @group getParam */
-  @Since("2.1.0")
-  def getPercentile: Double = $(percentile)
-
-  /**
-   * The highest p-value for features to be kept.
-   * Only applicable when selectorType = "fpr".
-   * Default value is 0.05.
-   * @group param
-   */
-  @Since("2.1.0")
-  final val fpr = new DoubleParam(this, "fpr", "The highest p-value for features to be kept.",
-    ParamValidators.inRange(0, 1))
-  setDefault(fpr -> 0.05)
-
-  /** @group getParam */
-  @Since("2.1.0")
-  def getFpr: Double = $(fpr)
-
-  /**
-   * The upper bound of the expected false discovery rate.
-   * Only applicable when selectorType = "fdr".
-   * Default value is 0.05.
-   * @group param
-   */
-  @Since("2.2.0")
-  final val fdr = new DoubleParam(this, "fdr",
-    "The upper bound of the expected false discovery rate.", ParamValidators.inRange(0, 1))
-  setDefault(fdr -> 0.05)
-
-  /** @group getParam */
-  def getFdr: Double = $(fdr)
-
-  /**
-   * The upper bound of the expected family-wise error rate.
-   * Only applicable when selectorType = "fwe".
-   * Default value is 0.05.
-   * @group param
-   */
-  @Since("2.2.0")
-  final val fwe = new DoubleParam(this, "fwe",
-    "The upper bound of the expected family-wise error rate.", ParamValidators.inRange(0, 1))
-  setDefault(fwe -> 0.05)
-
-  /** @group getParam */
-  def getFwe: Double = $(fwe)
-
-  /**
-   * The selector type of the ChisqSelector.
-   * Supported options: "numTopFeatures" (default), "percentile", "fpr", "fdr", "fwe".
-   * @group param
-   */
-  @Since("2.1.0")
-  final val selectorType = new Param[String](this, "selectorType",
-    "The selector type. Supported options: numTopFeatures, percentile, fpr, fdr, fwe",
-    ParamValidators.inArray(Array("numTopFeatures", "percentile", "fpr", "fdr",
-      "fwe")))
-  setDefault(selectorType -> "numTopFeatures")
-
-  /** @group getParam */
-  @Since("2.1.0")
-  def getSelectorType: String = $(selectorType)
-}
 
 /**
  * Chi-Squared feature selection, which selects categorical features to use for predicting a
@@ -153,106 +46,72 @@ private[feature] trait ChiSqSelectorParams extends Params
  */
 @Since("1.6.0")
 final class ChiSqSelector @Since("1.6.0") (@Since("1.6.0") override val uid: String)
-  extends Estimator[ChiSqSelectorModel] with ChiSqSelectorParams with DefaultParamsWritable {
+  extends Selector[ChiSqSelectorModel] {
 
   @Since("1.6.0")
   def this() = this(Identifiable.randomUID("chiSqSelector"))
 
   /** @group setParam */
   @Since("1.6.0")
-  def setNumTopFeatures(value: Int): this.type = set(numTopFeatures, value)
+  override def setNumTopFeatures(value: Int): this.type = super.setNumTopFeatures(value)
 
   /** @group setParam */
   @Since("2.1.0")
-  def setPercentile(value: Double): this.type = set(percentile, value)
+  override def setPercentile(value: Double): this.type = super.setPercentile(value)
 
   /** @group setParam */
   @Since("2.1.0")
-  def setFpr(value: Double): this.type = set(fpr, value)
+  override def setFpr(value: Double): this.type = super.setFpr(value)
 
   /** @group setParam */
   @Since("2.2.0")
-  def setFdr(value: Double): this.type = set(fdr, value)
+  override def setFdr(value: Double): this.type = super.setFdr(value)
 
   /** @group setParam */
   @Since("2.2.0")
-  def setFwe(value: Double): this.type = set(fwe, value)
+  override def setFwe(value: Double): this.type = super.setFwe(value)
 
   /** @group setParam */
   @Since("2.1.0")
-  def setSelectorType(value: String): this.type = set(selectorType, value)
+  override def setSelectorType(value: String): this.type = super.setSelectorType(value)
 
   /** @group setParam */
   @Since("1.6.0")
-  def setFeaturesCol(value: String): this.type = set(featuresCol, value)
+  override def setFeaturesCol(value: String): this.type = super.setFeaturesCol(value)
 
   /** @group setParam */
   @Since("1.6.0")
-  def setOutputCol(value: String): this.type = set(outputCol, value)
+  override def setOutputCol(value: String): this.type = super.setOutputCol(value)
 
   /** @group setParam */
   @Since("1.6.0")
-  def setLabelCol(value: String): this.type = set(labelCol, value)
+  override def setLabelCol(value: String): this.type = super.setLabelCol(value)
 
-  @Since("2.0.0")
+  /**
+   * get the SelectionTestResult for every feature against the label
+   */
+  protected[this] override def getSelectionTestResult(df: DataFrame): DataFrame = {
+    ChiSquareTest.test(df, getFeaturesCol, getLabelCol, true)
+  }
+
+  /**
+   * Create a new instance of concrete SelectorModel.
+   * @param indices The indices of the selected features
+   * @return A new SelectorModel instance
+   */
+  protected[this] def createSelectorModel(
+      uid: String,
+      indices: Array[Int]): ChiSqSelectorModel = {
+    new ChiSqSelectorModel(uid, indices)
+  }
+
+  @Since("1.6.0")
   override def fit(dataset: Dataset[_]): ChiSqSelectorModel = {
-    transformSchema(dataset.schema, logging = true)
-    val spark = dataset.sparkSession
-    import spark.implicits._
-
-    val numFeatures = MetadataUtils.getNumFeatures(dataset, $(featuresCol))
-    val resultDF = ChiSquareTest.test(dataset.toDF, $(featuresCol), $(labelCol), true)
-
-    def getTopIndices(k: Int): Array[Int] = {
-      resultDF.sort("pValue", "featureIndex")
-        .select("featureIndex")
-        .limit(k)
-        .as[Int]
-        .collect()
-    }
-
-    val indices = $(selectorType) match {
-      case "numTopFeatures" =>
-        getTopIndices($(numTopFeatures))
-      case "percentile" =>
-        getTopIndices((numFeatures * getPercentile).toInt)
-      case "fpr" =>
-        resultDF.select("featureIndex")
-          .where(col("pValue") < $(fpr))
-          .as[Int].collect()
-      case "fdr" =>
-        // This uses the Benjamini-Hochberg procedure.
-        // https://en.wikipedia.org/wiki/False_discovery_rate#Benjamini.E2.80.93Hochberg_procedure
-        val f = $(fdr) / numFeatures
-        val maxIndex = resultDF.sort("pValue", "featureIndex")
-          .select("pValue")
-          .as[Double].rdd
-          .zipWithIndex
-          .flatMap { case (pValue, index) =>
-            if (pValue <= f * (index + 1)) {
-              Iterator.single(index.toInt)
-            } else Iterator.empty
-          }.fold(-1)(math.max)
-        if (maxIndex >= 0) {
-          getTopIndices(maxIndex + 1)
-        } else Array.emptyIntArray
-      case "fwe" =>
-        resultDF.select("featureIndex")
-          .where(col("pValue") < $(fwe) / numFeatures)
-          .as[Int].collect()
-      case errorType =>
-        throw new IllegalStateException(s"Unknown Selector Type: $errorType")
-    }
-
-    copyValues(new ChiSqSelectorModel(uid, indices.sorted).setParent(this))
+     super.fit(dataset)
   }
 
   @Since("1.6.0")
-  override def transformSchema(schema: StructType): StructType = {
-    SchemaUtils.checkColumnType(schema, $(featuresCol), new VectorUDT)
-    SchemaUtils.checkNumericType(schema, $(labelCol))
-    SchemaUtils.appendColumn(schema, $(outputCol), new VectorUDT)
-  }
+  override def transformSchema(schema: StructType): StructType = super.transformSchema(schema)
 
   @Since("1.6.0")
   override def copy(extra: ParamMap): ChiSqSelector = defaultCopy(extra)
@@ -271,68 +130,23 @@ object ChiSqSelector extends DefaultParamsReadable[ChiSqSelector] {
 @Since("1.6.0")
 final class ChiSqSelectorModel private[ml] (
     @Since("1.6.0") override val uid: String,
-    @Since("3.1.0") val selectedFeatures: Array[Int])
-  extends Model[ChiSqSelectorModel] with ChiSqSelectorParams with MLWritable {
+    @Since("3.1.0") override val selectedFeatures: Array[Int])
+  extends SelectorModel[ChiSqSelectorModel](uid, selectedFeatures)  {
 
   import ChiSqSelectorModel._
 
-  if (selectedFeatures.length >= 2) {
-    require(selectedFeatures.sliding(2).forall(l => l(0) < l(1)),
-      "Index should be strictly increasing.")
-  }
+  override protected def isNumericAttribute = false
 
   /** @group setParam */
   @Since("1.6.0")
-  def setFeaturesCol(value: String): this.type = set(featuresCol, value)
+  override def setFeaturesCol(value: String): this.type = super.setFeaturesCol(value)
 
   /** @group setParam */
   @Since("1.6.0")
-  def setOutputCol(value: String): this.type = set(outputCol, value)
-
-  @Since("2.0.0")
-  override def transform(dataset: Dataset[_]): DataFrame = {
-    val outputSchema = transformSchema(dataset.schema, logging = true)
-
-    val newSize = selectedFeatures.length
-    val func = { vector: Vector =>
-      vector match {
-        case SparseVector(_, indices, values) =>
-          val (newIndices, newValues) = compressSparse(indices, values)
-          Vectors.sparse(newSize, newIndices, newValues)
-        case DenseVector(values) =>
-          Vectors.dense(selectedFeatures.map(values))
-        case other =>
-          throw new UnsupportedOperationException(
-            s"Only sparse and dense vectors are supported but got ${other.getClass}.")
-      }
-    }
-
-    val transformer = udf(func)
-    dataset.withColumn($(outputCol), transformer(col($(featuresCol))),
-      outputSchema($(outputCol)).metadata)
-  }
+  override def setOutputCol(value: String): this.type = super.setOutputCol(value)
 
   @Since("1.6.0")
-  override def transformSchema(schema: StructType): StructType = {
-    SchemaUtils.checkColumnType(schema, $(featuresCol), new VectorUDT)
-    val newField = prepOutputField(schema)
-    SchemaUtils.appendColumn(schema, newField)
-  }
-
-  /**
-   * Prepare the output column field, including per-feature metadata.
-   */
-  private def prepOutputField(schema: StructType): StructField = {
-    val selector = selectedFeatures.toSet
-    val origAttrGroup = AttributeGroup.fromStructField(schema($(featuresCol)))
-    val featureAttributes: Array[Attribute] = if (origAttrGroup.attributes.nonEmpty) {
-      origAttrGroup.attributes.get.zipWithIndex.filter(x => selector.contains(x._2)).map(_._1)
-    } else {
-      Array.fill[Attribute](selector.size)(NominalAttribute.defaultAttr)
-    }
-    val newAttributeGroup = new AttributeGroup($(outputCol), featureAttributes)
-    newAttributeGroup.toStructField()
-  }
+  override def transformSchema(schema: StructType): StructType = super.transformSchema(schema)
 
   @Since("1.6.0")
   override def copy(extra: ParamMap): ChiSqSelectorModel = {
@@ -347,41 +161,11 @@ final class ChiSqSelectorModel private[ml] (
   override def toString: String = {
     s"ChiSqSelectorModel: uid=$uid, numSelectedFeatures=${selectedFeatures.length}"
   }
-
-  private[spark] def compressSparse(
-      indices: Array[Int],
-      values: Array[Double]): (Array[Int], Array[Double]) = {
-    val newValues = new ArrayBuilder.ofDouble
-    val newIndices = new ArrayBuilder.ofInt
-    var i = 0
-    var j = 0
-    var indicesIdx = 0
-    var filterIndicesIdx = 0
-    while (i < indices.length && j < selectedFeatures.length) {
-      indicesIdx = indices(i)
-      filterIndicesIdx = selectedFeatures(j)
-      if (indicesIdx == filterIndicesIdx) {
-        newIndices += j
-        newValues += values(i)
-        j += 1
-        i += 1
-      } else {
-        if (indicesIdx > filterIndicesIdx) {
-          j += 1
-        } else {
-          i += 1
-        }
-      }
-    }
-    // TODO: Sparse representation might be ineffective if (newSize ~= newValues.size)
-    (newIndices.result(), newValues.result())
-  }
 }
 
 @Since("1.6.0")
 object ChiSqSelectorModel extends MLReadable[ChiSqSelectorModel] {
 
-  private[ChiSqSelectorModel]
   class ChiSqSelectorModelWriter(instance: ChiSqSelectorModel) extends MLWriter {
 
     private case class Data(selectedFeatures: Seq[Int])
